feat(on screen help): Implementation
- Implemented help icon button
