+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-08-17T12:13:03.557Z\n"
-"PO-Revision-Date: 2023-08-17T12:13:03.557Z\n"
+"POT-Creation-Date: 2023-08-17T13:05:10.512Z\n"
+"PO-Revision-Date: 2023-08-17T13:05:10.512Z\n"
 
 msgid "Action Configuration"
 msgstr "Action Configuration"
@@ -174,6 +174,15 @@ msgstr "Configure who can access this configuration and metadata associated"
 msgid "Configure sharing"
 msgstr "Configure sharing"
 
+msgid ""
+"This page allows you to configure basic details of the application. "
+"Settings like default dimensions, planning and tracking configuration, and "
+"access can be configured here."
+msgstr ""
+"This page allows you to configure basic details of the application. "
+"Settings like default dimensions, planning and tracking configuration, and "
+"access can be configured here."
+
 msgid "Organisation unit access"
 msgstr "Organisation unit access"
 
@@ -254,6 +263,67 @@ msgstr ""
 "Here you can set the default organisation unit. This will be the "
 "pre-selected organisation unit whenever the application starts"
 
+msgid ""
+"Planning frequency is the frequency at which you expect to plan your "
+"activities. When selected period selection will only be limited to periods "
+"of the frequency. When creating actions/activities, their start and end "
+"dates should be within the selected planning period."
+msgstr ""
+"Planning frequency is the frequency at which you expect to plan your "
+"activities. When selected period selection will only be limited to periods "
+"of the frequency. When creating actions/activities, their start and end "
+"dates should be within the selected planning period."
+
+msgid "Note"
+msgstr "Note"
+
+msgid "The planning frequency should be greater or equal to the tracking period."
+msgstr "The planning frequency should be greater or equal to the tracking period."
+
+msgid ""
+"Tracking frequency is the frequency at which you expect to track your "
+"activities. It controls how frequent you are supposed to provide an "
+"activity status. While on tracking mode, this will control the type of "
+"periods in the tracking columns."
+msgstr ""
+"Tracking frequency is the frequency at which you expect to track your "
+"activities. It controls how frequent you are supposed to provide an "
+"activity status. While on tracking mode, this will control the type of "
+"periods in the tracking columns."
+
+msgid "The tracking frequency should be less or equal to the planning period."
+msgstr "The tracking frequency should be less or equal to the planning period."
+
+msgid ""
+"Here you can set the default period. This will be the pre-selected period "
+"whenever the application starts"
+msgstr ""
+"Here you can set the default period. This will be the pre-selected period "
+"whenever the application starts"
+
+msgid "You can change the year for the default period here."
+msgstr "You can change the year for the default period here."
+
+msgid "Reset changes"
+msgstr "Reset changes"
+
+msgid ""
+"Click here if you want to redo all changes on all configuration. This will "
+"be disabled when there are no changes to redo."
+msgstr ""
+"Click here if you want to redo all changes on all configuration. This will "
+"be disabled when there are no changes to redo."
+
+msgid "Save changes"
+msgstr "Save changes"
+
+msgid ""
+"Click here to save all changes done. This will be disabled when there are "
+"no changes to save"
+msgstr ""
+"Click here to save all changes done. This will be disabled when there are "
+"no changes to save"
+
 msgid "Error saving changes"
 msgstr "Error saving changes"
 
@@ -265,9 +335,6 @@ msgstr "Reset"
 
 msgid "Updating metadata..."
 msgstr "Updating metadata..."
-
-msgid "Save changes"
-msgstr "Save changes"
 
 msgid "Back"
 msgstr "Back"
