+++ src/modules/Config/components/General/General.tsx
@@ -3,8 +3,8 @@ import i18n from '@dhis2/d2-i18n'
 import { Divider } from '@dhis2/ui'
 import { OrganisationUnitConfig } from './components/OrganisationUnitConfig'
 import { PeriodConfig } from './components/PeriodConfig'
-import { GeneralSteps } from './docs/steps'
-import { HelpButton } from '../../../../shared/components/HelpButton'
+import { GeneralSteps, PeriodSteps } from './docs/steps'
+import { HelpButton, HelpIcon } from '../../../../shared/components/HelpButton'
 
 export function General () {
 
@@ -21,7 +21,10 @@ export function General () {
                 <div className="gap-16 column">
                     <h3 className="m-0">{i18n.t('Organisation Units')}</h3>
                     <OrganisationUnitConfig/>
-                    <h3 className="m-0">{i18n.t('Period')}</h3>
+                    <div className="row gap-8 align-center">
+                        <h3 className="m-0">{i18n.t('Period')}</h3>
+                        <HelpIcon steps={PeriodSteps} key="period-config-steps"/>
+                    </div>
                     <PeriodConfig/>
                 </div>
             </div>
